REFACTOR ProductObject
Move product info to a dataobject. Product page has_one Product object. Allows products to be attached to multiple Product pages across categories.
